refactor(cli): extract quarantine boundary
